+++ keyring/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2023-06-06
+
+### Added
+
+- Added `Entry::find` function.
+- Exposed `keyring::Error` as `KeyringError`.
+
 ## [0.0.2] - 2023-06-06
 
 ### Added